CloudflareDurable.Application (cloudflare_durable v0.2.0)

Application interface for Cloudflare Durable Objects.

This module provides an interface for managing state in a Durable Object.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

delete(pid, key)

@spec delete(pid(), String.t()) :: :ok | {:error, term()}

Deletes a value from the Durable Object state.

Parameters

pid - PID of the Application process

key - Key to delete

Returns

:ok - Successfully deleted the value

{:error, reason} - Failed to delete the value

get(pid, key)

@spec get(pid(), String.t()) :: {:ok, term()} | {:error, term()}

Gets a value from the Durable Object state.

Parameters

pid - PID of the Application process

key - Key to get

Returns

{:ok, value} - Successfully got the value

{:error, reason} - Failed to get the value

get_all(pid, keys)

@spec get_all(pid(), [String.t()]) :: {:ok, map()} | {:error, term()}

Gets multiple values from the Durable Object state.

Parameters

pid - PID of the Application process

keys - List of keys to get

Returns

{:ok, map} - Successfully got the values

{:error, reason} - Failed to get the values

put(pid, key, value)

@spec put(pid(), String.t(), term()) :: :ok | {:error, term()}

Puts a value in the Durable Object state.

Parameters

pid - PID of the Application process

key - Key to put

value - Value to put

Returns

:ok - Successfully put the value

{:error, reason} - Failed to put the value

put_all(pid, map)

@spec put_all(pid(), map()) :: :ok | {:error, term()}

Puts multiple values in the Durable Object state.

Parameters

pid - PID of the Application process

map - Map of key-value pairs to put

Returns

:ok - Successfully put the values

{:error, reason} - Failed to put the values

start_link(opts)

@spec start_link(keyword()) :: GenServer.on_start()

Starts a new Application process.

Parameters

opts - Options for the Application:

:url - URL of the Cloudflare Worker (required)

:id_type - Type of ID to use (:string or :name, default: :string)

:object_name - Object name to use (for :name ID type)

:retry_options - Retry options for HTTP requests

Returns

{:ok, pid} - Successfully started the Application

{:error, reason} - Failed to start the Application

stop(pid)

@spec stop(pid()) :: :ok

Stops the Application process.

Parameters

pid - PID of the Application process

Returns

:ok - Successfully stopped the process

update(pid, key, fun)

@spec update(pid(), String.t(), (term() -> term())) ::
  {:ok, term()} | {:error, term()}

Updates a value in the Durable Object state using a function.

Parameters

pid - PID of the Application process

key - Key to update

fun - Function to apply to the current value

Returns

{:ok, new_value} - Successfully updated the value

{:error, reason} - Failed to update the value