CloudflareDurable.Benchmark.Utils (cloudflare_durable v0.2.0)

Utility functions for benchmarks.

This module provides common utility functions used by the benchmark modules.

Functions

ensure_dir(path)

@spec ensure_dir(String.t()) :: :ok

Creates a directory if it doesn't exist.

Parameters

path - The path to create

Returns

:ok - The directory was created or already exists

format_bytes(bytes)

@spec format_bytes(float()) :: String.t()

Formats a byte size in the most appropriate unit.

Parameters

bytes - The size in bytes

Returns

String.t() - A formatted size string

format_time(time_ns)

@spec format_time(float()) :: String.t()

Formats a time value in the most appropriate unit.

Parameters

time_ns - The time in nanoseconds

Returns

String.t() - A formatted time string

random_id()

@spec random_id() :: String.t()

Generates a random ID for use in benchmarks.

Returns

String.t() - A random string ID

random_string(length)

@spec random_string(non_neg_integer()) :: String.t()

Generates a random string of the specified length.

Parameters

length - The length of the random string to generate

Returns

String.t() - A random string of the specified length

summarize_results(results)

@spec summarize_results(map()) :: String.t()

Summarizes benchmark results for the console.

Parameters

results - The benchmark results

Returns

String.t() - A string summarizing the results

timestamp()

@spec timestamp() :: String.t()

Gets a timestamp for use in filenames.

Returns

String.t() - A timestamp string in the format "YYYYMMDD_HHMMSS"