CloudflareDurable.Client (cloudflare_durable v0.2.0)

Client for interacting with Cloudflare Durable Objects.

This module provides functions for communicating with Cloudflare Workers and Durable Objects, allowing applications to leverage edge-located, stateful storage and processing.

Types

client_opts()

@type client_opts() :: [
  worker_url: String.t(),
  api_token: String.t(),
  name: atom(),
  finch_pool_size: pos_integer(),
  finch_pool_count: pos_integer()
]

connection_opts()

@type connection_opts() :: [
  url: String.t(),
  auto_reconnect: boolean(),
  backoff_initial: non_neg_integer(),
  backoff_max: non_neg_integer(),
  subscriber: pid() | nil
]

error_reason()

@type error_reason() ::
  :network_error
  | :invalid_response
  | :server_error
  | :not_found
  | atom()
  | String.t()

http_body()

@type http_body() :: String.t() | nil

http_headers()

@type http_headers() :: [{String.t(), String.t()}]

http_method()

@type http_method() :: :get | :post | :put | :delete

http_response()

@type http_response() :: %{
  status: http_status(),
  body: map() | nil,
  headers: http_headers()
}

http_status()

@type http_status() :: non_neg_integer()

method_name()

@type method_name() :: String.t()

object_id()

@type object_id() :: String.t()

t()

@type t() :: module()

Functions

call_method(object_id, method, params, opts \\ [])

@spec call_method(object_id(), method_name(), map(), keyword()) ::
  {:ok, map()} | {:error, error_reason()}

Calls a method on a Durable Object.

Parameters

object_id - ID of the Durable Object to call

method - Method to call on the Durable Object

params - Parameters to pass to the method

opts - Optional parameters:

:worker_url - Override the default worker URL

Returns

{:ok, response} - Successfully called method on Durable Object

{:error, reason} - Failed to call method on Durable Object

delete_state(object_id, key, opts \\ [])

@spec delete_state(object_id(), String.t(), keyword()) ::
  {:ok, map()} | {:error, error_reason()}

Deletes a key from a Durable Object's state.

Parameters

object_id - ID of the Durable Object to update

key - Key to delete

opts - Optional parameters:

:worker_url - Override the default worker URL

Returns

{:ok, response} - Successfully deleted key

{:error, reason} - Failed to delete key

get_namespace_object(client, namespace, name, opts \\ [])

@spec get_namespace_object(t(), String.t(), String.t(), keyword()) ::
  {:ok, String.t()} | {:error, error_reason()}

Gets a namespace object ID from a namespace and name.

Parameters

client - Client module

namespace - Durable Object namespace

name - Name within the namespace

opts - Optional parameters:

:worker_url - Override the default worker URL

Returns

{:ok, object_id} - Successfully got object ID

{:error, reason} - Failed to get object ID

get_state(object_id, key \\ nil, opts \\ [])

@spec get_state(object_id(), String.t() | nil, keyword()) ::
  {:ok, map()} | {:error, error_reason()}

Gets the state of a Durable Object.

Parameters

object_id - ID of the Durable Object to get state from

key - Optional specific key to get (if nil, gets all state)

opts - Optional parameters:

:worker_url - Override the default worker URL

Returns

{:ok, state} - Successfully retrieved state

{:error, reason} - Failed to retrieve state

initialize(object_id, data, opts \\ [])

@spec initialize(object_id(), map(), keyword()) ::
  {:ok, map()} | {:error, error_reason()}

Initializes a new Durable Object instance.

Parameters

object_id - ID of the Durable Object to initialize

data - Initial data to store in the Durable Object

opts - Optional parameters:

:worker_url - Override the default worker URL

Returns

{:ok, response} - Successfully initialized Durable Object

{:error, reason} - Failed to initialize Durable Object

open_websocket(object_id, opts \\ [])

@spec open_websocket(
  object_id(),
  keyword()
) :: {:ok, pid()} | {:error, error_reason()}

Opens a WebSocket connection to a Durable Object.

Parameters

object_id - ID of the Durable Object to connect to

opts - Optional parameters:

:worker_url - Override the default worker URL

:auto_reconnect - Whether to automatically reconnect on disconnect (default: true)

:backoff_initial - Initial backoff time in ms (default: 500)

:backoff_max - Maximum backoff time in ms (default: 30000)

Returns

{:ok, pid} - Successfully opened WebSocket connection

{:error, reason} - Failed to open WebSocket connection

update_state(object_id, key, value, opts \\ [])

@spec update_state(object_id(), String.t(), any(), keyword()) ::
  {:ok, map()} | {:error, error_reason()}

Updates the state of a Durable Object.

Parameters

object_id - ID of the Durable Object to update

key - Key to update

value - Value to set

opts - Optional parameters:

:worker_url - Override the default worker URL

Returns

{:ok, response} - Successfully updated state

{:error, reason} - Failed to update state

websocket_connect(client, object_id, path, opts \\ [])

@spec websocket_connect(t(), String.t(), String.t(), keyword()) ::
  {:ok, pid()} | {:error, error_reason()}

Establishes a WebSocket connection to a Durable Object.

Parameters

client - Client module

object_id - ID of the Durable Object to connect to

path - Path to connect to

opts - Connection options:

:subscriber - PID to receive WebSocket messages

:auto_reconnect - Whether to automatically reconnect on disconnect (default: true)

:backoff_initial - Initial backoff time in ms (default: 500)

:backoff_max - Maximum backoff time in ms (default: 30000)

Returns

{:ok, pid} - Successfully established WebSocket connection

{:error, reason} - Failed to establish WebSocket connection